API improvements

+++ Balta/6.0/ToDo/Controllers/HomeController.cs
@@ -8,59 +8,63 @@ namespace ToDo.Controllers
     public class HomeController : ControllerBase
     {
         [HttpGet("/")]
-        public List<TodoModels> Get([FromServices]AppDbContext context)
-        {
-            return context.Todos.ToList();
-        }
+        public IActionResult Get([FromServices] AppDbContext context)
+            => Ok(context.Todos.ToList());
         
         [HttpGet("/{id:int}")]
-        public TodoModels GetById(
+        public IActionResult GetById(
             [FromRoute] int id,
             [FromServices]AppDbContext context)
         {
-            return context.Todos.FirstOrDefault(x => x.Id == id);
+            var todos = context.Todos.FirstOrDefault(x => x.Id == id);
+            if (todos == null)
+                return NotFound();
+
+            return Ok(todos);
         }
         
         [HttpPost("/")]
-        public TodoModels Post(
+        public IActionResult Post(
             [FromBody]TodoModels todo,
             [FromServices] AppDbContext context)
         {
             context.Todos.Add(todo);
             context.SaveChanges();
 
-            return todo;
+            return Created($"/{todo.Id}", todo);
         }
         
         [HttpPut("/")]
-        public TodoModels Put(
+        public IActionResult Put(
             [FromRoute]int id,
             [FromBody]TodoModels todo,
             [FromServices] AppDbContext context)
         {
             var model = context.Todos.FirstOrDefault(x => x.Id == id);
             if (model == null)
-                return todo;
+                return NotFound();
 
             model.Title = todo.Title;
             model.Done = todo.Done;
 
             context.Todos.Update(model);
             context.SaveChanges();
-            return model;
+            return Ok(model);
         }
         
         [HttpDelete("/")]
-        public TodoModels Delete(
+        public IActionResult Delete(
             [FromRoute]int id,
             [FromBody]TodoModels todo,
             [FromServices] AppDbContext context)
         {
             var model = context.Todos.FirstOrDefault(x => x.Id == id);
+            if (model == null)
+                return NotFound();
 
             context.Todos.Remove(model);
             context.SaveChanges();
-            return model;
+            return Ok(model);
         }
     }
